+++ src/components/plot/outline-dropdown-menu.tsx
@@ -134,7 +134,7 @@ export function OutlineDropdownMenu({
 
       <DropdownMenuContent
         align={align}
-        variant="content"
+        //variant="content"
         className="flex flex-col"
       >
         {color0.showThemeColors && (